// applications/cpq-maps/src/app/cpq-map/models/availability-detail-response.ts
export interface AvailabilityDetailResponse {
  inhomeGeoId: number;
  productCode: string;
  productName: string;
  ihw: string;
  geoProfileId: string;
  geocode: string;
  zip: string;
  profileTypeAbbr: string;
  wrapGeoProfileId: string;
  wrapName: string;
  city: string;
  sdmGeoProfileId: number;
  sdmName: string;
  pricingMarketGeoProfileId: number;
  pricingMarketName: string;
  primaryDeliveryMethodAbbr: string;
  subDeliveryMethodAbbr: string;
  distributionMethod: string;
  ownerGroup: string;
  ihday: string;
  sharedHhcPossible: string;
  sharedHhcScheduled: string;
  isAvailable: 0 | 1;
  pobHhc: number;
  isPobOnly: 0 | 1;
  isWeekly: 0 | 1;
  ownerAccountNumber: number;
  ownerAccountName: string;
  isAlternateDelivery: 0 | 1;
  alternateDeliveryAccountNumber: number;
  alternateDeliveryAccountName: string;
  productionDepartmentId: number;
  productionDepartmentName: string;
  subProductionDepartmentId: number;
  subProductionDepartmentName: string;
  pagePositions: string;
  atzAnalysis: 'Y' | 'N';
  firstIhd: string;
  secondIhd: string;
}